"revise"

+++ src/gemm_fpga.h
@@ -26,7 +26,7 @@ extern void gemm_nn(int M, int N, int K, float ALPHA,
 extern void gemm_fpga_finalize();
 
 #ifdef OPENEXR
-void gemm_nn_fpga_half(int M, int N, int K, half ALPHA, 
+void gemm_nn_fpga_half(int M, int N, int K, float ALPHA, 
                     half *A, int lda, 
                     half *B, int ldb,
                     half *C, int ldc);
